+++ original/list/linkedlist.h
@@ -6,17 +6,17 @@
 typedef struct _tagMTL_LinkedList
 {
 	void *type;
-	int length;
+	size_t length;
 	void *head, *tail;
 } *MTL_LinkedList;
 
 MTL_LinkedList new_MTL_LinkedList();
 void MTL_LinkedList_release(MTL_LinkedList self);
 
-int MTL_LinkedList_length(MTL_LinkedList self);
-bool MTL_LinkedList_set(MTL_LinkedList self, int idx, void *value);
-bool MTL_LinkedList_get(MTL_LinkedList self, int idx, void **value);
-bool MTL_LinkedList_append(MTL_LinkedList self, int *idx, void *value);
+size_t MTL_LinkedList_length(MTL_LinkedList self);
+bool MTL_LinkedList_set(MTL_LinkedList self, size_t idx, void *value);
+bool MTL_LinkedList_get(MTL_LinkedList self, size_t idx, void **value);
+bool MTL_LinkedList_append(MTL_LinkedList self, size_t *idx, void *value);
 bool MTL_LinkedList_iterate(MTL_LinkedList self, bool (*func)(void *value));
 
 #endif
